api/requirements.txt
Django==3.2.7
djangorestframework==3.12.4
django-cors-headers==3.8.0
django-redis==5.2.0
drf-spectacular==0.26.5
drf-spectacular-sidecar==2023.12.1
factory-boy==3.3.0
msgpack==1.0.3
parameterized==0.9.0
psycopg2-binary==2.9.1
uWSGI==2.0.20
Werkzeug==1.0.1